aider: The database schema has been updated to change the `chatrooms` table so that the `id` field is a text field instead of an integer.
# Aider chat conversation:

USER: Please modify the system so that room ids are strings/text fields instead of integers. Just edit one set of endpoint, db field, openapi spec section at a time. Don't try to update all of them at once please. Once you have the first edit complete let me know and I will prompt you to continue. Thanks very much!

ASSISTANT: Certainly! We'll start by updating the database schema to change the `chatrooms` table so that the `id` field is a text field instead of an integer. This will involve modifying the `database.js` file.

Here's the change for the `chatrooms` table:

```javascript
database.js
<<<<<<< SEARCH
db.run(`CREATE TABLE IF NOT EXISTS chatrooms (
    id INTEGER PRIMARY KEY AUTOINCREMENT,
=======
db.run(`CREATE TABLE IF NOT EXISTS chatrooms (
    id TEXT PRIMARY KEY,
>>>>>>> REPLACE
    name TEXT NOT NULL
)`);
```

This change will make the `id` field of the `chatrooms` table a text field and remove the `AUTOINCREMENT` property since it's not applicable to text fields.

Please note that this change will require existing data to be migrated if there is any, as the type of the primary key is being changed. If there is existing data, a migration strategy will need to be planned and executed.

Once you confirm this change, we can proceed to update the corresponding API endpoint and OpenAPI specification.