aider: Reverted the HTTP methods for joining a chat room and sending a message back to GET.
# Aider chat conversation:

USER: Unfortunately we cannot use POST at all due to an integration restriction with another party. Please make sure they are all GET

ASSISTANT: Understood. To comply with the integration restriction, I will revert the HTTP methods for the endpoints in `app.js` to GET. Here are the changes that need to be made:

1. Change the HTTP method for joining a chat room back to GET.
2. Change the HTTP method for sending a message back to GET.
3. Ensure that the endpoint for retrieving messages remains a GET request.

Let's start with the changes to `app.js`:

```javascript
app.js
<<<<<<< SEARCH
// API endpoint to join a chat room
app.post('/chatrooms/:roomId/users', (req, res) => {
=======
// API endpoint to join a chat room
app.get('/chatrooms/:roomId/users', (req, res) => {
>>>>>>> REPLACE

<<<<<<< SEARCH
// API endpoint to send a message to a chat room
app.post('/chatrooms/:roomId/messages', (req, res) => {
=======
// API endpoint to send a message to a chat room
app.get('/chatrooms/:roomId/messages', (req, res) => {
>>>>>>> REPLACE
```

These changes will revert the HTTP methods for the specified endpoints back to GET. Please note that using GET for actions such as sending a message is not a best practice, as GET requests should be idempotent and not change the state of the server. However, if this is a requirement due to integration constraints, the above changes will implement it as requested.